python/requirements.txt
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.4
chromadb>=0.5.11
semantic-router>=0.5.3
semantic-chunkers>=0.5.3
pydantic>=2.8.0
